caseiq-backend/requirements/base.txt
# Core Django
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.10
pgvector==0.3.2

# Async Tasks
celery==5.4.0
redis==5.1.0
django-celery-results==2.5.1

# PDF Processing
pdfplumber==0.11.4
PyMuPDF==1.24.10

# AI & Embeddings
groq==0.11.0
sentence-transformers==3.1.1
numpy==1.26.4
scikit-learn==1.5.2

# Utilities
Pillow==10.4.0
python-slugify==8.0.4